update devcontainer and function blueprints
