+++ .devcontainer/Dockerfile
@@ -1,8 +1,1 @@
-ARG IMAGE=bullseye
-FROM --platform=amd64 mcr.microsoft.com/devcontainers/${IMAGE}
-RUN curl https://packages.microsoft.com/keys/microsoft.asc | gpg --dearmor > microsoft.gpg \
-    && mv microsoft.gpg /etc/apt/trusted.gpg.d/microsoft.gpg \
-    && sh -c 'echo "deb [arch=amd64] https://packages.microsoft.com/debian/$(lsb_release -rs | cut -d'.' -f 1)/prod $(lsb_release -cs) main" > /etc/apt/sources.list.d/dotnetdev.list' \
-    && apt-get update && apt-get install -y xdg-utils \
-    && apt-get update && apt-get install -y azure-functions-core-tools-4 \
-    && apt-get clean -y && rm -rf /var/lib/apt/lists/*
+FROM mcr.microsoft.com/devcontainers/dotnet:0-7.0-bullseye
