+++ .devcontainer/devcontainer.json
@@ -1,41 +1,40 @@
 {
-    "name": "Azure Developer CLI",
-    "build": {
-        "dockerfile": "Dockerfile",
-        "args": {
-            "IMAGE": "dotnet:6.0"
-        }
-    },
-    "features": {
-        "ghcr.io/devcontainers/features/docker-in-docker:2": {},
-        "ghcr.io/devcontainers/features/node:1": {},
-        "ghcr.io/azure/azure-dev/azd": {
-            "version": "latest"
-        },
-        "ghcr.io/devcontainers/features/python:1": {},
-        "ghcr.io/azure/azure-dev/azd:0": {},
-        "ghcr.io/devcontainers/features/azure-cli:1": {}
-    },
-    "customizations": {
-        "vscode": {
-            "extensions": [
-                "GitHub.vscode-github-actions",
-                "ms-azuretools.azure-dev",
-                "ms-azuretools.vscode-azurefunctions",
-                "ms-azuretools.vscode-bicep",
-                "ms-azuretools.vscode-docker",
-                "ms-dotnettools.csharp",
-                "ms-dotnettools.vscode-dotnet-runtime",
-                "ms-vscode.vscode-node-azure-pack"
-            ]
-        }
-    },
-    "forwardPorts": [
-        7071
-    ],
-    "postCreateCommand": "",
-    "remoteUser": "vscode",
-    "hostRequirements": {
-        "memory": "8gb"
+  "name": "Azure Functions (.NET)",
+  "dockerComposeFile": "docker-compose.yml",
+  "service": "app",
+  "workspaceFolder": "/workspaces/${localWorkspaceFolderBasename}",
+  "forwardPorts": [
+    7071,
+    10000,
+    10001,
+    10002
+  ],
+  "otherPortsAttributes": {
+    "onAutoForward": "ignore"
+  },
+  "features": {
+    "ghcr.io/devcontainers/features/azure-cli:1": {},
+	"ghcr.io/jlaundry/devcontainer-features/azure-functions-core-tools:1": {},
+	"ghcr.io/devcontainers/features/python:1": {
+    "version": "3.11"
+  },
+  "ghcr.io/azure/azure-dev/azd:0": {},
+  "ghcr.io/devcontainers/features/node:1": {}
+  },
+  "customizations": {
+    "vscode": {
+      "extensions": [
+        "ms-azuretools.vscode-azurefunctions",
+        "ms-dotnettools.csdevkit",
+        "ms-python.python",
+        "humao.rest-client",
+        "azurite.azurite",
+        "ms-toolsai.jupyter"
+      ],
+      "settings": {
+        "azureFunctions.projectSubpath": "semantic-kernel-azure-function"
+      }
     }
+  },
+  "postCreateCommand": "npm install -g azurite"
 }
